refactor(entity): 将 Admin 类中的 id 类型从 Long 改为 Integer
- 修改了 Admin 类中的 id 字段类型，以更好地匹配数据库表结构
- 这个改动可能会影响使用该字段的相关方法和接口

+++ backend/src/main/java/cn/hmck/entity/Admin.java
@@ -11,7 +11,7 @@ import java.time.LocalDateTime;
 @TableName("er_admin")
 public class Admin implements Serializable {
     @TableId(type = IdType.AUTO)
-    private Long id;  // 管理员ID
+    private Integer id;  // 管理员ID
 
     private String username;  // 管理员用户名
